refactor: Add viewset votes logic to helper view
