+++ posts/api/v1/views/nested_post_viewset.py
@@ -1,12 +1,10 @@
 """
-API V1: Topic ViewSet
+API V1: Post ViewSet
 """
 ###
 # Libraries
 ###
-from rest_framework import viewsets, permissions
-from rest_framework.decorators import action
-from rest_framework.response import Response
+from rest_framework import viewsets
 
 from topics.models import Topic
 from posts.models import Post
@@ -14,20 +12,14 @@ from posts.api.v1.serializers import (
     PostListSerializer,
     PostDetailSerializer,
 )
-from helpers.permissions import IsOwnerOrReadOnly
+from helpers.views import VotesViewSetMixin
 from helpers.services import VotesService, DeleteService
 
 
 ###
 # ViewSet
 ###
-class NestedPostViewSet(viewsets.ModelViewSet):
-    def get_permissions(self):
-        self.permission_classes = [permissions.IsAuthenticatedOrReadOnly]
-        if not self.action.endswith('vote'):
-            self.permission_classes.append(IsOwnerOrReadOnly)
-        return super().get_permissions()
-
+class NestedPostViewSet(VotesViewSetMixin, viewsets.ModelViewSet):
     def get_queryset(self):
         topic_slug = self.kwargs.get('topic_slug')
         qs = Post.objects.filter(topic__slug=topic_slug)
@@ -47,27 +39,3 @@ class NestedPostViewSet(viewsets.ModelViewSet):
 
     def perform_destroy(self, instance):
         DeleteService(instance).update_author_to_deleted()
-
-    @action(detail=True, methods=['patch'], url_path='up-vote')
-    def up_vote(self, request, *args, **kwargs):
-        obj = self.get_object()
-        service = VotesService(obj=obj, user=request.user)
-        if obj.is_up_voted:
-            obj = service.remove_up_vote()
-        else:
-            obj = service.add_up_vote()
-
-        serializer = self.get_serializer(obj)
-        return Response(serializer.data)
-
-    @action(detail=True, methods=['patch'], url_path='down-vote')
-    def down_vote(self, request, *args, **kwargs):
-        obj = self.get_object()
-        service = VotesService(obj=obj, user=request.user)
-        if obj.is_down_voted:
-            obj = service.remove_down_vote()
-        else:
-            obj = service.add_down_vote()
-
-        serializer = self.get_serializer(obj)
-        return Response(serializer.data)
